src/requirements.txt
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.15
langchain-openai==0.1.7
langgraph
Flask
Flask-Migrate
Flask-CORS
python-dotenv
psycopg2-binary
web3